+++ src/frontend/js/widgets/Dashboard/components/DashboardListAvatar/index.tsx
@@ -1,9 +1,20 @@
+import { StringHelper } from 'utils/StringHelper';
 import { DashboardAvatar, DashboardAvatarProps } from '../DashboardAvatar';
 
-const DashboardListAvatar = (props: DashboardAvatarProps) => {
+interface DashboardListAvatarProps extends DashboardAvatarProps {
+  saturation?: number;
+  lightness?: number;
+}
+
+const DashboardListAvatar = ({
+  saturation,
+  lightness,
+  ...avatarProps
+}: DashboardListAvatarProps) => {
+  const backgroundColor = StringHelper.toColor(avatarProps.title, saturation, lightness);
   return (
     <div className="dashboard-list-avatar__container">
-      <DashboardAvatar {...props} />
+      <DashboardAvatar {...avatarProps} backgroundColor={backgroundColor} />
     </div>
   );
 };